# tsoa/api/dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Prisma.UserCreateWithoutPostsInput": {
				"properties": {
					"updatedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email"
				],
				"type": "object"
			},
			"Prisma.StringFieldUpdateOperationsInput": {
				"properties": {
					"set": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"Prisma.NullableStringFieldUpdateOperationsInput": {
				"properties": {
					"set": {
						"type": "string",
						"nullable": true
					}
				},
				"type": "object"
			},
			"Prisma.DateTimeFieldUpdateOperationsInput": {
				"properties": {
					"set": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"type": "object"
			},
			"Prisma.UserUpdateWithoutPostsInput": {
				"properties": {
					"updatedAt": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Prisma.DateTimeFieldUpdateOperationsInput"
							},
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"createdAt": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Prisma.DateTimeFieldUpdateOperationsInput"
							},
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"password": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Prisma.StringFieldUpdateOperationsInput"
							},
							{
								"type": "string"
							}
						]
					},
					"name": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Prisma.NullableStringFieldUpdateOperationsInput"
							},
							{
								"type": "string"
							}
						],
						"nullable": true
					},
					"email": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Prisma.StringFieldUpdateOperationsInput"
							},
							{
								"type": "string"
							}
						]
					},
					"id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Prisma.StringFieldUpdateOperationsInput"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": ""
	},
	"paths": {
		"/api/prisma/{id}": {
			"post": {
				"operationId": "Post",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"properties": {
												"posts": {
													"items": {
														"properties": {
															"authorId": {
																"type": "string"
															},
															"updatedAt": {
																"type": "string",
																"format": "date-time"
															},
															"createdAt": {
																"type": "string",
																"format": "date-time"
															},
															"content": {
																"type": "string"
															},
															"title": {
																"type": "string"
															},
															"id": {
																"type": "string"
															}
														},
														"required": [
															"authorId",
															"updatedAt",
															"createdAt",
															"content",
															"title",
															"id"
														],
														"type": "object"
													},
													"type": "array"
												}
											},
											"required": [
												"posts"
											],
											"type": "object"
										},
										{
											"properties": {
												"updatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"password": {
													"type": "string"
												},
												"name": {
													"type": "string",
													"nullable": true
												},
												"email": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"updatedAt",
												"createdAt",
												"password",
												"name",
												"email",
												"id"
											],
											"type": "object"
										},
										{
											"properties": {
												"uwu": {
													"properties": {
														"uwu": {
															"properties": {
																"uwu": {
																	"properties": {
																		"uwu": {
																			"type": "string",
																			"enum": [
																				"owo"
																			],
																			"nullable": false
																		}
																	},
																	"required": [
																		"uwu"
																	],
																	"type": "object"
																}
															},
															"required": [
																"uwu"
															],
															"type": "object"
														}
													},
													"required": [
														"uwu"
													],
													"type": "object"
												}
											},
											"required": [
												"uwu"
											],
											"type": "object"
										}
									]
								},
								"examples": {
									"Example 1": {},
									"Example 2": {}
								}
							}
						}
					},
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Prisma.UserCreateWithoutPostsInput"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "Patch",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"properties": {
												"author": {
													"properties": {
														"updatedAt": {
															"type": "string",
															"format": "date-time"
														},
														"createdAt": {
															"type": "string",
															"format": "date-time"
														},
														"password": {
															"type": "string"
														},
														"name": {
															"type": "string",
															"nullable": true
														},
														"email": {
															"type": "string"
														},
														"id": {
															"type": "string"
														}
													},
													"required": [
														"updatedAt",
														"createdAt",
														"password",
														"name",
														"email",
														"id"
													],
													"type": "object"
												}
											},
											"required": [
												"author"
											],
											"type": "object"
										},
										{
											"properties": {
												"authorId": {
													"type": "string"
												},
												"updatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"content": {
													"type": "string"
												},
												"title": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"authorId",
												"updatedAt",
												"createdAt",
												"content",
												"title",
												"id"
											],
											"type": "object"
										}
									]
								},
								"examples": {
									"Example 1": {},
									"Example 2": {},
									"Example 3": {},
									"Example 4": {},
									"Example 5": {},
									"Example 6": {},
									"Example 7": {},
									"Example 8": {},
									"Example 9": {}
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"enum": [
										"Not OK"
									],
									"nullable": false
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "number",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Prisma.UserUpdateWithoutPostsInput"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}